+++ CHANGELOG.md
@@ -5,7 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+##Added
+- Add Srm login with access and refresh token (mutation autenticarUsuario)[(3)]
+- Add Srm user register (mutation nuevoUsuario) [(3)]
+- Add base project[(3)]
+    
 ### DevOps
+- Add node_modules to volumen [()]
 - Add docker support [(1)]
 - New folder structure [(1)]
 - Add CI/CD from codeship [(1)]
@@ -13,3 +19,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [Unreleased]: https://bitbucket.org/cumplo/srm-apollo-server/branches/compare/devel..#diff
 
 [(1)]: https://bitbucket.org/cumplo/srm-apollo-server/pull-requests/1/
+[(3)]: https://bitbucket.org/cumplo/srm-apollo-server/pull-requests/3/